fix: NUR API connect failure
If the NUR API has not been able to connect or has timed out multiple times, the service is still bound yet API is agnostic of this which will keep the service alive and bluetooth connection alive and no way to recover from API side. In this case, just kill the UART service and start it again.